requirements.txt
requests>=2.25.1
Pillow>=8.0.0
Flask>=2.0.0
Flask-SocketIO>=5.0.0
python-socketio>=5.0.0
eventlet>=0.33.0
gunicorn>=20.1.0